fix compiling issue after renaming
